[FIRRTL] Update NLAs in ModuleInliner pass
Fix the ModuleInliner pass to update non-local annotations (NLAs) in
response to inlining/flattening.  This handles arbitrary NLAs whose
root (first module in the NLA) is an ancestor of the inlined/flattened
module, is the inlined/flattened module, or is a child of the
inlined/flattened module.

Broadly, this commit adds logic for updating NLAs during module inlining
for the following situations:

  1. Paths in an NLA are dropped due to inlining.
  2. An NLA is truncated at a flattened module when flattened.
  3. Combinations of (1) and (2) do the right thing.
  4. NLAs whose root module is inlined are copied to create multiple,
     new NLAs rooted on all modules that instantiate the inlined module.
  5. Any result of (1) and (2) that makes an NLA local results in the
     NLA leaf being promoted to a local annotation.
  6. Any unused NLAs are garbage collected.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>